requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.9.0
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
python-dotenv==1.0.1
Werkzeug==3.0.2
